Cambio a singular de la funcion renderizarProducto() y mover la misma funcion al archivo productos.js
